Change condition for setting limit on bg traffic sched ...
Before: if the special sched didn't run recently
Now: if we aren't currently counting cells

Before if the special sched was sending out a trickle of leftover data
after a measurement finished, this would affect the regular sched.

Now as soon as we are told a measurement is over, we go back to allowing
the regular sched to send as much as it needs.