Have target relay skip decryption on echo circs ...
I think at one point we wanted it to perform the decryption since it
needs to do it for regular traffic, but we seem to be moving away from
that line of thinking.

This simply moves the speical case of "I just got a cell on an echo
circuit and will now send it right back" to **before** the decryption.