Stop loop from becoming infinite if circuit issue
Recently Tor clients have started hitting 100% CPU and going
unresponsive to kill signals other than -9 after an experiment
completes.

Turns out the problem is in circuit_send_speedtest_cells() calling
relay_send_command_from_edge() calling circuit_package_relay_cell()
(which actually appends the cell to the circuit queue), which never
returns anything other than 0. Sometimes are circuits are marked for
close at the end of an experiment, causing circuit_package_relay_cell()
to return 0 early without actually sending the cell, causing the while
loop in circuit_send_speedtest_cells() to be infinite.

As a fix I landed on checking if the current time is is past the end of
the experiment stop time, and if so, to exit the while loop. Other
possible solutions I thought of:

- only sending some maximum number of cells per call to
circuit_send_speedtest_cells(), but I wasn't sure if that would Just
Work (i.e. would it reliably get called again) and didn't want to figure
that out, if not.

- not letting the number of echo cells we've sent grow too much more
than the number of echo cells we've recveived. This would be at odds
with the High/Low water circuit queue limits at the target and could
lead to surprising behavior (e.g., it's pretty low at the client side,
the target keeps increasing its water marks but never sees a change in
throughput or number of buffererd cells).