Try to ensure relay is told about msm starting before ...
the measurement clients start sending their first echo cells. It's the
job of the bg "measurement" client to tell the relay. Previous changes
on the relay side that made the relay force close circuits/conns on
which it sees echo cells when not in a measurement, there is a race
condition. That is: if the bg client needs to get its signal cell to the
target before too many of the measurers' first cells get there.
Otherwise ~all or sufficiently many of the measurement circuits will get
closed and the measurement won't be useful.